// types/technologies.ts
export enum Languages {
    TypeScript = "TypeScript",
    JavaScript = "JavaScript",
    NodeJs = "Node.js",
    React = "React",
    NextJs = "Next.js",
    RestApi = "REST API",
    TailwindCSS = "TailwindCSS",
    ChakraUI = "Chakra-UI",
    GraphQL = "GraphQL",
    Vercel = "Vercel",
    Storybook = "storybook",
    StyledComponents = "Styled Components",
}